+++ src/App.tsx
@@ -1,9 +1,9 @@
 
 import { Routes, Route } from 'react-router-dom';
-import StartPage from './pages/StartPage';
-import { GridMenu } from './pages/GridMenu';
-import Chat from './pages/ChatPage';
-import { UtilityBar } from './components/UtilityBar';
+import StartPage from './pages/StartPage/StartPage';
+import { GridMenu } from './pages/GridMenu/GridMenu';
+import Chat from './pages/ChatPage/ChatPage';
+import { UtilityBar } from './components/UtilityBar/UtilityBar';
 import { DarkThemeProvider } from './darkthemeprovider/DarkThemeProvider';
 
 function App() {
