+++ src/components/BackButton/BackButton.css
@@ -1,0 +1,13 @@
+/* BackButton.css */
+.back-button {
+  display: flex;
+  align-items: center;
+  gap: 0.25rem; 
+  padding: 0.5rem 0.75rem; 
+  border-radius: 9999px;
+}
+
+.back-button-icon {
+  width: 1.5rem; 
+  height: 1.5rem;
+}
