+++ src/components/ColorModeSwitch/ColorModeSwitch.tsx
@@ -1,12 +1,11 @@
 import { Button, Tooltip } from "@digdir/designsystemet-react";
 import { SunIcon, MoonIcon } from '@navikt/aksel-icons';
-import { ColorMode, useColorMode } from '../lib/settings';
-
+import { ColorMode, useColorMode } from '../../lib/settings';
+import './ColorModeSwitch.css';
 
 export default function ColorModeSwitch() {
   const [colorMode, setColorMode] = useColorMode();
 
-  // Fallback to system on first load — but only toggle light/dark manually
   const toggleColorMode = () => {
     const newMode = colorMode === 'dark' ? ColorMode.Light : ColorMode.Dark;
     setColorMode(newMode);
@@ -15,12 +14,11 @@ export default function ColorModeSwitch() {
   const icon = colorMode === 'dark' ? <SunIcon /> : <MoonIcon />;
   const tooltipText = colorMode === 'dark' ? "Switch to light mode" : "Switch to dark mode";
 
-
   return (
     <Tooltip content={tooltipText} placement="bottom">
       <Button
         variant="tertiary"
-        className="bg-transparent text-[color:var(--ds-color-main-text)] p-0 text-lg hover:bg-[color:var(--ds-color-neutral-surface-hover)] hover:text-[color:var(--ds-color-main-text)] focus:outline-none border-none outline-none"
+        className="color-mode-button"
         onClick={toggleColorMode}
       >
         {icon}
@@ -28,3 +26,4 @@ export default function ColorModeSwitch() {
     </Tooltip>
   );
 }
+
