+++ src/components/DropdownMenu/DropdownMenu.css
@@ -1,0 +1,51 @@
+/* DropDownMenu.css */
+
+.dropdown-container {
+  position: relative;
+  width: 100%;
+  max-width: 20rem; 
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.dropdown-trigger {
+  width: 100%;
+  margin-top: 0.25rem; 
+  background-color: transparent;
+  border: none;
+  outline: none;
+  padding: 0.5rem 1rem; 
+  border-radius: 0.375rem; 
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 0.5rem;
+  transition: background-color 0.2s ease;
+}
+
+.dropdown-trigger:hover,
+.dropdown-trigger:focus,
+.dropdown-trigger:active {
+  background-color: var(--ds-color-neutral-surface-hover);
+}
+
+.dropdown-list {
+  width: 100%;
+  max-height: 15rem; 
+  overflow-y: auto;
+  box-shadow: 0 10px 15px rgba(0, 0, 0, 0.1); 
+  border-radius: 0.375rem; 
+  background-color: white;
+  z-index: 50;
+}
+
+.dropdown-button {
+  width: 100%;
+  text-align: left;
+  padding: 0.5rem 1rem; 
+  transition: background-color 0.2s ease;
+}
+
+.dropdown-button:hover {
+  background-color: var(--ds-color-neutral-surface-hover);
+}
