+++ src/components/DropdownMenu/DropdownMenu.tsx
@@ -2,17 +2,17 @@ import { Dropdown } from '@digdir/designsystemet-react';
 import { ChevronDownIcon } from '@navikt/aksel-icons';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { useState, useMemo } from 'react';
-import { slugify } from '../utils/slugify';
+import { slugify } from '../../utils/slugify';
+import './DropDownMenu.css';
 
 interface DropDownMenuProps {
   solutions: string[];
 }
 
-// Format fallback slug to title case
 function formatSlug(slug: string): string {
   return slug
-    .replace(/-/g, ' ') // Replace hyphens with spaces
-    .replace(/\b\w/g, (char) => char.toUpperCase()); // Capitalize each word
+    .replace(/-/g, ' ')
+    .replace(/\b\w/g, (char) => char.toUpperCase());
 }
 
 export const DropDownMenu = ({ solutions }: DropDownMenuProps) => {
@@ -20,12 +20,10 @@ export const DropDownMenu = ({ solutions }: DropDownMenuProps) => {
   const navigate = useNavigate();
   const location = useLocation();
 
-  // Find base path (brukerstøtte or servicedesk)
   const pathParts = location.pathname.split('/').filter(Boolean);
   const basePath = pathParts[0] ?? '';
-  const currentSlug = (pathParts[1] ?? '').toLowerCase(); // lowercased to match slugify
+  const currentSlug = (pathParts[1] ?? '').toLowerCase();
 
-  // Find selected solution that matches current slug
   const selected = useMemo(() => {
     return solutions.find((solution) => slugify(solution) === currentSlug);
   }, [currentSlug, solutions]);
@@ -39,23 +37,23 @@ export const DropDownMenu = ({ solutions }: DropDownMenuProps) => {
   };
 
   return (
-    <div className="relative w-full max-w-xs mx-auto">
+    <div className="dropdown-container">
       <Dropdown.TriggerContext>
         <Dropdown.Trigger
           onClick={() => setOpen(!open)}
-          className="w-full mt-1 bg-transparent border-none outline-none hover:bg-[var(--ds-color-neutral-surface-hover)] focus:bg-[var(--ds-color-neutral-surface-hover)] active:bg-[var(--ds-color-neutral-surface-hover)] rounded px-4 py-2 flex items-center justify-between gap-2"
+          className="dropdown-trigger"
         >
           {selected ?? (currentSlug ? formatSlug(currentSlug) : 'Velg løsning')}
           <ChevronDownIcon aria-hidden />
         </Dropdown.Trigger>
 
         <Dropdown open={open} onClose={() => setOpen(false)}>
-          <Dropdown.List className="w-full max-h-60 overflow-y-auto shadow-lg rounded-md bg-white z-50">
+          <Dropdown.List className="dropdown-list">
             {solutions.map((solution) => (
               <Dropdown.Button
                 key={solution}
                 onClick={() => handleSelect(solution)}
-                className="px-4 py-2 w-full text-left hover:bg-[var(--ds-color-neutral-surface-hover)]"
+                className="dropdown-button"
               >
                 {solution}
               </Dropdown.Button>
