+++ src/components/InputField/InputField.css
@@ -1,0 +1,50 @@
+.input-field-container {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  max-width: none;
+
+}
+
+.input-field-input {
+  width: 100%;
+  height: 100%;
+  padding: 1rem 3.5rem 1rem 1.25rem;
+  border: none;
+  border-radius: 20px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  outline: none;
+  transition: box-shadow 0.2s ease;
+}
+
+.input-field-input:focus {
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.15);
+}
+
+.input-field-button {
+  position: absolute;
+  top: 50%;
+  right: 0.5rem;
+  transform: translateY(-50%);
+  height: 70%;
+  aspect-ratio: 1 / 1;
+  border-radius: 9999px;
+  background-color: transparent;
+  padding: 0;
+  margin: 0;
+  font-size: 1.125rem;
+  color: var(--ds-color-neutral-text-default);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: color 0.2s ease;
+}
+
+.input-field-button:hover {
+  color: var(--ds-color-neutral-text-subtle);
+}
+
+.input-field-icon {
+  width: 1.25rem;
+  height: 1.25rem;
+}
