+++ src/components/InputField/InputField.tsx
@@ -1,6 +1,7 @@
 import { Button, Input } from '@digdir/designsystemet-react';
 import { PaperplaneIcon } from '@navikt/aksel-icons';
 import { useState } from 'react';
+import './InputField.css';
 
 type InputFieldProps = {
   onSend: (message: string) => void;
@@ -16,20 +17,20 @@ export function InputField({ onSend }: InputFieldProps) {
   };
 
   return (
-    <div className="relative w-full h-full">
+    <div className="input-field-container">
       <Input
         placeholder="Spør et spørsmål"
         value={inputValue}
         onChange={(e) => setInputValue(e.target.value)}
         onKeyDown={(e) => e.key === 'Enter' && handleSend()}
-        className="w-full h-full rounded-[20px] pr-14 px-5 py-4 border-none shadow-md focus:outline-none focus:shadow-lg"
+        className="input-field-input"
       />
       <Button
         variant="primary"
         onClick={handleSend}
-        className="absolute top-1/2 right-2 -translate-y-1/2 h-[70%] aspect-square rounded-full bg-transparent p-0 m-0 text-lg text-[var(--ds-color-neutral-text-default)] hover:text-[var(--ds-color-neutral-text-subtle)] flex items-center justify-center"
+        className="input-field-button"
       >
-        <PaperplaneIcon className="w-5 h-5" />
+        <PaperplaneIcon className="input-field-icon" />
       </Button>
     </div>
   );
