+++ src/components/NavigationButton/NavigationButton.tsx
@@ -1,5 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 import { ArrowRightIcon } from '@navikt/aksel-icons';
+import './NavigationButton.css';
 
 type NavigationButtonProps = {
   label: string;
@@ -8,18 +9,16 @@ type NavigationButtonProps = {
   className?: string;
 };
 
- export function NavigationButton({ label, to, icon, className }: NavigationButtonProps) {
+export function NavigationButton({ label, to, icon, className }: NavigationButtonProps) {
   const navigate = useNavigate();
 
   return (
     <button
       onClick={() => navigate(to)}
-      className={`w-full text-[var(--ds-color-text-default)] font-medium px-4 py-3 rounded-full shadow-sm
-        btn-theme
-        transition flex items-center justify-between ${className || ''}`}
+      className={`navigation-button btn-theme ${className || ''}`}
     >
       <span>{label}</span>
-      {icon ?? <ArrowRightIcon className="w-5 h-5" />}
+      {icon ?? <ArrowRightIcon className="navigation-button-icon" />}
     </button>
   );
 }
