+++ src/components/SolutionCard/SolutionCard.css
@@ -1,0 +1,32 @@
+/* SolutionCard.css */
+
+.solution-card {
+  background-color: var(--ds-color-interactive-primary);
+  width: 100%;
+  aspect-ratio: 1 / 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  border-radius: 0.375rem; /* rounded-md */
+  border: 1px solid var(--ds-color-interactive-primary); /* border-primary */
+  transition: border-color 0.2s ease;
+}
+
+.solution-card:hover {
+  border-color: var(--ds-color-interactive-primary); /* hover:border-primary */
+}
+
+.solution-card-block {
+  padding-left: 0.25rem;
+  padding-right: 0.25rem;
+}
+
+.solution-card-title {
+  font-size: 11px;
+  font-weight: 500;
+  line-height: 1.375;
+  color: var(--ds-color-interactive-primary); /* text-primary */
+  word-break: break-word;
+  hyphens: auto;
+}
