+++ src/components/UtilityBar/UtilityBar.css
@@ -1,0 +1,30 @@
+/* UtilityBar.css */
+
+.utility-bar {
+  position: fixed;
+  top: 0.5rem; /* top-2 */
+  right: 0.5rem; /* right-2 */
+  z-index: 100;
+  display: flex;
+  align-items: center;
+  gap: 0.125rem; /* gap-0.5 */
+}
+
+.utility-button {
+  background-color: transparent;
+  color: var(--ds-color-main-text);
+  padding: 0;
+  font-size: 1.125rem; /* text-lg */
+  border: none;
+  outline: none;
+  transition: background-color 0.2s ease, color 0.2s ease;
+}
+
+.utility-button:hover {
+  background-color: var(--ds-color-neutral-surface-hover);
+  color: var(--ds-color-main-text);
+}
+
+.utility-button:focus {
+  outline: none;
+}
