+++ src/pages/ChatPage/ChatPage.css
@@ -1,0 +1,101 @@
+.chat-container {
+  position: relative;
+  background-color: var(--ds-color-neutral-background-subtle);
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.chat-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: start;
+  width: 100%;
+  padding: 1rem 1rem 0;
+}
+
+.chat-header-left {
+  display: flex;
+  flex-direction: row;
+  gap: 0.5rem;
+}
+
+.chat-body {
+  flex: 1;
+  width: 100%;
+  overflow-y: auto;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.chat-footer {
+  width: 100%;
+  padding: 1rem;
+  display: flex;
+  align-items: center;
+  box-sizing: border-box;
+}
+
+/* Wrapper for Textarea og knapp */
+.chat-input-wrapper {
+  position: relative;
+  width: 100%;
+  flex-grow: 1;
+}
+
+/* Viktig: Overstyr wrapper fra Digdir-komponenten */
+.chat-input-wrapper .ds-textarea__wrapper {
+  width: 100% !important;
+  max-width: 100% !important;
+  box-sizing: border-box;
+}
+
+/* Viktig: Overstyr Digdir sin Textarea direkte */
+.chat-input.ds-input {
+  width: 350px !important;
+  max-width: 100% !important;
+  min-height: 3rem;
+  max-height: 8rem;
+  resize: none;
+  overflow-y: auto;
+  padding: 0.75rem 3rem 0.75rem 1rem !important;
+  border-radius: 20px;
+  border: none;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  outline: none;
+  font-family: inherit;
+  font-size: 1rem;
+  box-sizing: border-box;
+  --dsc-input-size: auto;
+  --dsc-input-padding: 0.75rem 3rem 0.75rem 1rem;
+}
+
+.chat-send-button {
+  position: absolute;
+  right: 0.75rem;
+  top: 50%;
+  transform: translateY(-50%);
+  height: 2rem;
+  width: 2rem;
+  border-radius: 9999px;
+  background-color: transparent;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: var(--ds-color-neutral-text-default);
+  border: none;
+  cursor: pointer;
+}
+
+.chat-send-button:hover {
+  color: var(--ds-color-neutral-text-subtle);
+}
+
+.chat-send-icon {
+  width: 1.25rem;
+  height: 1.25rem;
+}
