+++ src/pages/GridMenu/GridMenu.css
@@ -1,0 +1,51 @@
+/* GridMenu.css */
+
+.grid-menu {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  min-height: 100vh;
+  padding: 1rem;
+  text-align: center;
+  background-color: #f2f3f5;
+}
+
+.dark .grid-menu {
+  background-color: var(--ds-color-brand3-background-tinted);
+}
+
+.grid-menu-header {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  margin-bottom: 1rem;
+}
+
+.grid-menu-title {
+  font-size: 0.75rem; /* text-xs */
+  font-weight: 500;   /* font-medium */
+  margin-top: 0.25rem;
+  color: var(--ds-color-text-default);
+}
+
+@media (min-width: 640px) {
+  .grid-menu-title {
+    font-size: 0.875rem; /* sm:text-sm */
+  }
+}
+
+.grid-menu-logo {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 2rem;
+}
+
+.grid-menu-content {
+  width: 100%;
+  max-width: 600px;
+  margin-left: auto;
+  margin-right: auto;
+}
