+++ src/pages/StartPage/StartPage.css
@@ -1,0 +1,53 @@
+/* StartPage.css */
+
+.start-page {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 2.5rem 1.5rem; /* py-10 px-6 */
+  background-color: #f2f3f5;
+}
+
+.dark .start-page {
+  background-color: var(--ds-color-brand3-background-tinted);
+}
+
+.start-page-content {
+  margin-top: 4rem; /* mt-16 */
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.start-page-inner {
+  width: 80%;
+  max-width: 28rem; /* max-w-md */
+  text-align: center;
+}
+
+.start-page-logo {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 2rem;
+}
+
+.start-page-heading {
+  font-size: 1.875rem; /* text-3xl */
+  font-weight: 300; /* font-light */
+  line-height: 1.375;
+  margin-bottom: 1.5rem;
+  color: #002c54;
+}
+
+.dark .start-page-heading {
+  color: var(--ds-color-text-default);
+}
+
+.start-page-buttons {
+  margin-top: 4rem; /* mt-16 */
+  display: flex;
+  flex-direction: column;
+  gap: 1rem; /* gap-4 */
+}
